requirements.txt
streamlit>=1.44.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
PyPDF2>=3.0.0
faiss-cpu>=1.11.0
langchain_google_genai>=2.0.0
httpx>=0.28.0
